[DM-25653] Refactor out businesses into separate files
Each business should be in a file, so split it up and make the
manager easily compatible with that.
